Moved propertyEligibilityValidationMock to helpers.js file
